Simplify EMU_RunInterrupts usage
Previously, the behavior of this function when detecting an interrupt
was to return the address of the interrupt handler routine. Instead, run
an execution loop until the handler returns. As this means the return
value will always be equivalent to the argument, return void instead.
The interrupt handler routines do not appear to care about their return
addresses, so remove the argument as well and use 0xDEADBEEF.

This removes a large number of (potentially costly) conditional branches
from ZZ_Native and significantly simplifies its control flow. Usage in
other functions such as CORE_VSync is also simplified.